Start refactoring BlockViewModel into a mixin.
