+++ rover/src/main/kotlin/io/rover/rover/ui/BlockViewModelFactory.kt
@@ -30,6 +30,9 @@ class BlockViewModelFactory(
                 measurementService,
                 BackgroundViewModel(block),
                 BorderViewModel(block)
+                // TODO: I would need to pass in some sort of measure-ator to blockviewmodel...
+                // actually, I could take the Text Concerns themselves out and put them into their own headless viewmodel, and have that implement
+                // the measure-ator interface :)
             )
             is ImageBlock -> ImageBlockViewModel(
                 block,
