+++ rover/src/main/kotlin/io/rover/rover/ui/viewmodels/BackgroundViewModel.kt
@@ -3,7 +3,7 @@ package io.rover.rover.ui.viewmodels
 import io.rover.rover.core.domain.Background
 import io.rover.rover.ui.views.asAndroidColor
 
-open class BackgroundViewModel(
+class BackgroundViewModel(
     val background: Background
 ): BackgroundViewModelInterface {
     override val backgroundColor: Int
