+++ rover/src/main/kotlin/io/rover/rover/ui/views/Interfaces.kt
@@ -39,9 +39,9 @@ interface ViewBorderInterface {
 }
 
 interface ViewTextInterface {
-    var textViewModel: TextBlockViewModelInterface?
+    var textBlockViewModel: TextBlockViewModelInterface?
 }
 
 interface ViewImageInterface {
-    var imageViewModel: ImageBlockViewModelInterface?
+    var imageBlockViewModel: ImageBlockViewModelInterface?
 }
